+++ bridge/opencensus/internal/oc2otel/span_context_test.go
@@ -6,6 +6,11 @@ package oc2otel
 import (
 	"testing"
 
+	"github.com/stretchr/testify/assert"
+	"go.opencensus.io/plugin/ochttp/propagation/tracecontext"
+
+	"go.opentelemetry.io/otel/bridge/opencensus/internal/otel2oc"
+
 	octrace "go.opencensus.io/trace"
 	"go.opencensus.io/trace/tracestate"
 
@@ -13,10 +18,21 @@ import (
 )
 
 func TestSpanContextConversion(t *testing.T) {
+	tsOc, _ := tracestate.New(nil,
+		tracestate.Entry{Key: "key1", Value: "value1"},
+		tracestate.Entry{Key: "key2", Value: "value2"},
+	)
+	tsOtel := trace.TraceState{}
+	tsOtel, _ = tsOtel.Insert("key2", "value2")
+	tsOtel, _ = tsOtel.Insert("key1", "value1")
+
+	httpFormatOc := &tracecontext.HTTPFormat{}
+
 	for _, tc := range []struct {
-		description string
-		input       octrace.SpanContext
-		expected    trace.SpanContext
+		description        string
+		input              octrace.SpanContext
+		expected           trace.SpanContext
+		expectedTracestate string
 	}{
 		{
 			description: "empty",
@@ -47,23 +63,32 @@ func TestSpanContextConversion(t *testing.T) {
 			}),
 		},
 		{
-			description: "trace state is ignored",
+			description: "trace state should be propagated",
 			input: octrace.SpanContext{
 				TraceID:    octrace.TraceID([16]byte{1}),
 				SpanID:     octrace.SpanID([8]byte{2}),
-				Tracestate: &tracestate.Tracestate{},
+				Tracestate: tsOc,
 			},
 			expected: trace.NewSpanContext(trace.SpanContextConfig{
-				TraceID: trace.TraceID([16]byte{1}),
-				SpanID:  trace.SpanID([8]byte{2}),
+				TraceID:    trace.TraceID([16]byte{1}),
+				SpanID:     trace.SpanID([8]byte{2}),
+				TraceState: tsOtel,
 			}),
+			expectedTracestate: "key1=value1,key2=value2",
 		},
 	} {
 		t.Run(tc.description, func(t *testing.T) {
 			output := SpanContext(tc.input)
-			if !output.Equal(tc.expected) {
-				t.Fatalf("Got %+v spancontext, expected %+v.", output, tc.expected)
-			}
+			assert.Equal(t, tc.expected, output)
+
+			// Ensure the otel tracestate and oc tracestate has the same header output
+			_, ts := httpFormatOc.SpanContextToHeaders(tc.input)
+			assert.Equal(t, tc.expectedTracestate, ts)
+			assert.Equal(t, tc.expectedTracestate, tc.expected.TraceState().String())
+
+			// The reverse conversion should yield the original input
+			input := otel2oc.SpanContext(output)
+			assert.Equal(t, tc.input, input)
 		})
 	}
 }